Read terragrunt version from .terragrunt-version file
The parallel tool tfswitch can read the required terraform
version from a `.terraform-version` file in addition to its
original `.tfswitchrc` file. This makes it compatible with a similar
tool, tfenv.

See: https://github.com/warrensbox/terraform-switcher#instead-of-a-tfswitchrc-file-a-terraform-version-file-may-be-used-for-compatibility-with-tfenv-and-other-tools-which-use-it

This commit adds a similar feature to tgswitch by reading a
`.terragrunt-version` file, which is compatible with tgenv.

See: https://github.com/cunymatthieu/tgenv#terragrunt-version